Ad system_networkState RPC (with cleanups)
