+++ packages/types/src/classes.ts
@@ -2,6 +2,6 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import * as Classes from './index';
+import * as Classes from './index.types';
 
 export type CodecTypes = keyof typeof Classes;
