+++ packages/types/src/index.types.ts
@@ -10,4 +10,5 @@
  */
 export { default as Metadata } from './Metadata';
 export * from './primitive';
+export * from './rpc';
 export * from './type';
