+++ packages/types/src/rpc/ChainProperties.spec.ts
