+++ packages/types/src/rpc/PendingExtrinsics.ts
@@ -1,0 +1,13 @@
+// Copyright 2017-2019 @polkadot/types authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+import Extrinsics from '../type/Extrinsics';
+
+/**
+ * @name PendingExtrinsics
+ * @description
+ * A list of pending [[Extrinsics]]
+ */
+export default class PendingExtrinsics extends Extrinsics {
+}
