+++ packages/types/src/rpc/index.ts
@@ -1,0 +1,20 @@
+// Copyright 2017-2019 @polkadot/types authors & contributors
+// This software may be modified and distributed under the terms
+// of the Apache-2.0 license. See the LICENSE file for details.
+
+// NOTE When adding any types here, we need to update the coumentation links as
+// well - <root>/docs/SUMMARY.md as well as ../README.md
+
+/**
+ * @summary Type definitions that are used in the rpc
+ */
+export { default as ChainProperties } from './ChainProperties';
+export { default as ExtrinsicStatus } from './ExtrinsicStatus';
+export { default as Health } from './Health';
+export { default as Json } from './Json';
+export { default as NetworkState } from './NetworkState';
+export { default as PeerInfo } from './PeerInfo';
+export { default as PendingExtrinsics } from './PendingExtrinsics';
+export { default as RuntimeVersion } from './RuntimeVersion';
+export { default as SignedBlock } from './SignedBlock';
+export { default as StorageChangeSet } from './StorageChangeSet';
