+++ packages/types/src/type/index.ts
@@ -20,7 +20,6 @@ export { default as Balance, BalanceOf } from './Balance';
 // export * from './Bft;
 export { default as Block } from './Block';
 export { default as BlockNumber } from './BlockNumber';
-export { default as ChainProperties } from './ChainProperties';
 export { default as CodeHash } from './CodeHash';
 export { default as Digest, DigestItem } from './Digest';
 export { default as Event } from './Event';
@@ -28,13 +27,11 @@ export { default as EventRecord } from './EventRecord';
 export { default as Extrinsic } from './Extrinsic';
 export { default as ExtrinsicEra } from './ExtrinsicEra';
 export { default as ExtrinsicSignature } from './ExtrinsicSignature';
-export { default as ExtrinsicStatus } from './ExtrinsicStatus';
 // NOTE Only used internally, exported as PendingExtrinsics
 // export { default as Extrinsics } from './Extrinsics';
 export { default as Gas } from './Gas';
 export { default as Hash } from './Hash';
 export { default as Header, HeaderExtended } from './Header';
-export { default as Health } from './Health';
 export { default as InherentOfflineReport } from './InherentOfflineReport';
 export { default as KeyValue } from './KeyValue';
 export { default as LockPeriods } from './LockPeriods';
@@ -46,8 +43,6 @@ export { default as IndexCompact } from './NonceCompact';
 export { default as Justification } from './Justification';
 export { default as Origin } from './Origin';
 export { default as ParaId } from './ParaId';
-export { default as PeerInfo } from './PeerInfo';
-export { default as PendingExtrinsics } from './PendingExtrinsics';
 export { default as Permill } from './Permill';
 export { default as Perbill } from './Perbill';
 export { default as PrefabWasmModule } from './PrefabWasmModule';
@@ -56,13 +51,11 @@ export { default as Proposal } from './Proposal';
 export { default as ProposalIndex } from './ProposalIndex';
 export { default as ReferendumIndex } from './ReferendumIndex';
 export { default as ReferendumInfo } from './ReferendumInfo';
-export { default as RuntimeVersion } from './RuntimeVersion';
 export { default as Schedule } from './Schedule';
 export { default as SeedOf } from './SeedOf';
 export { default as SessionKey } from './SessionKey';
 export { default as Signature, Ed25519Signature, Sr25519Signature } from './Signature';
-export { default as SignedBlock } from './SignedBlock';
-export { default as StorageChangeSet } from './StorageChangeSet';
+
 export { default as StoredPendingChange } from './StoredPendingChange';
 export { default as ValidatorPrefs } from './ValidatorPrefs';
 export { default as VestingSchedule } from './VestingSchedule';
